+++ composeApp/src/wasmJsMain/kotlin/tech/kotlinlang/permission/ui/camera/CameraScreen.wasmJs.kt
@@ -5,9 +5,11 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.LocalTextStyle
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
+import kotlinx.browser.window
 
 @Composable
 actual fun CameraScreen() {