+++ permission/src/wasmJsMain/kotlin/tech/kotlinlang/permission/location/WasmJsLocationHelper.kt
@@ -1,0 +1,7 @@
+package tech.kotlinlang.permission.location
+
+class WasmJsLocationHelper: LocationHelper {
+    override suspend fun fetchLastKnownLocation(): LocationRequestResult {
+        TODO("Not yet implemented")
+    }
+}